Moved object array to manager.js and displayed each DJ and their time by selecting each Producer
