+++ HTML/Manager/Manager.html
@@ -5,7 +5,6 @@
     <link rel="stylesheet" href="../../CSS/Manager/manager-design.css">
     <script type = "module" src = "../../Javascript/Manager/manager.js"></script>
     <script  src = "../../Javascript/Manager/manager-timeslots.js" ></script>
-
     <style>
         .pop-up:taget {
             display:block
@@ -28,9 +27,9 @@
             <header> Producers </header>
             <div class = "producer-list">
                 <ul>
-                    <li>PD1</li>
-                    <li>PD2</li>
-                    <li>PD3</li>
+                    <li>Producer1</li>
+                    <li>Producer2</li>
+                    <li>Producer3</li>
                 </ul
             </div>
         </section>
