--- Javascript/Manager/manager-timeslots.js
@@ -1,16 +1,0 @@
-let timeSlots = {
-    "Producer1": [
-        {name : "John Doe", date: "3/10/2024", start: "1:00pm", end: "3:00pm"},
-        {name : "Jane Doe", date: "3/11/2024", start: "4:00pm", end: "5:00pm"},
-        {name : "Willow Smith", date: "3/12/2024", start: "6:00pm", end: "7:00pm"},
-    ],
-    "Producer2": [
-        {name : "Bruce Wayne", date: "3/15/2024", start: "8:00pm", end: "9:00pm"},
-        {name : "Frederick Barbarossa", date: "3/16/2024", start: "10:00pm", end: "11:00pm"},
-    ],
-    "Producer3": [
-        {name : "Catherine de'Medici", date: "3/17/2024", start: "12:00am", end: "1:00am"},
-    ]
-}
-
-module.exports = timeSlots
